Refactor: Move storage struct into separate files
