+++ openraft/src/core/sm/handle.rs
@@ -4,12 +4,12 @@ use crate::async_runtime::MpscUnboundedSender;
 use crate::async_runtime::MpscUnboundedWeakSender;
 use crate::async_runtime::SendError;
 use crate::core::sm;
+use crate::storage::Snapshot;
 use crate::type_config::alias::JoinHandleOf;
 use crate::type_config::alias::MpscUnboundedSenderOf;
 use crate::type_config::alias::MpscUnboundedWeakSenderOf;
 use crate::type_config::TypeConfigExt;
 use crate::RaftTypeConfig;
-use crate::Snapshot;
 
 /// State machine worker handle for sending command to it.
 pub(crate) struct Handle<C>
