+++ openraft/src/engine/handler/snapshot_handler/mod.rs
@@ -5,9 +5,9 @@ use crate::display_ext::DisplayOptionExt;
 use crate::engine::Command;
 use crate::engine::EngineOutput;
 use crate::raft_state::LogStateReader;
+use crate::storage::SnapshotMeta;
 use crate::RaftState;
 use crate::RaftTypeConfig;
-use crate::SnapshotMeta;
 
 #[cfg(test)]
 mod trigger_snapshot_test;
