+++ openraft/src/lib.rs
@@ -76,6 +76,12 @@ pub use type_config::async_runtime;
 pub use type_config::async_runtime::tokio_impls::TokioRuntime;
 pub use type_config::AsyncRuntime;
 
+pub use self::storage::LogState;
+pub use self::storage::RaftLogReader;
+pub use self::storage::RaftSnapshotBuilder;
+pub use self::storage::Snapshot;
+pub use self::storage::SnapshotMeta;
+pub use self::storage::StorageHelper;
 pub use crate::base::OptionalSend;
 pub use crate::base::OptionalSerde;
 pub use crate::base::OptionalSync;
@@ -109,12 +115,6 @@ pub use crate::raft_state::MembershipState;
 pub use crate::raft_state::RaftState;
 pub use crate::raft_types::SnapshotId;
 pub use crate::raft_types::SnapshotSegmentId;
-pub use crate::storage::LogState;
-pub use crate::storage::RaftLogReader;
-pub use crate::storage::RaftSnapshotBuilder;
-pub use crate::storage::Snapshot;
-pub use crate::storage::SnapshotMeta;
-pub use crate::storage::StorageHelper;
 pub use crate::storage_error::ErrorSubject;
 pub use crate::storage_error::ErrorVerb;
 pub use crate::storage_error::StorageError;
