+++ tests/tests/client_api/t16_with_state_machine.rs
@@ -4,6 +4,8 @@ use anyhow::Result;
 use maplit::btreeset;
 use openraft::error::Fatal;
 use openraft::storage::RaftStateMachine;
+use openraft::storage::Snapshot;
+use openraft::storage::SnapshotMeta;
 use openraft::testing::log_id;
 use openraft::Config;
 use openraft::Entry;
@@ -11,8 +13,6 @@ use openraft::LogId;
 use openraft::OptionalSend;
 use openraft::RaftSnapshotBuilder;
 use openraft::RaftTypeConfig;
-use openraft::Snapshot;
-use openraft::SnapshotMeta;
 use openraft::StorageError;
 use openraft::StoredMembership;
 use openraft_memstore::ClientResponse;
