+++ tests/tests/fixtures/mod.rs
@@ -46,6 +46,7 @@ use openraft::raft::VoteRequest;
 use openraft::raft::VoteResponse;
 use openraft::storage::RaftLogStorage;
 use openraft::storage::RaftStateMachine;
+use openraft::storage::Snapshot;
 use openraft::Config;
 use openraft::LogId;
 use openraft::LogIdOptionExt;
@@ -58,7 +59,6 @@ use openraft::RaftMetrics;
 use openraft::RaftState;
 use openraft::RaftTypeConfig;
 use openraft::ServerState;
-use openraft::Snapshot;
 use openraft::Vote;
 use openraft_memstore::ClientRequest;
 use openraft_memstore::ClientResponse;
