+++ tests/tests/snapshot_streaming/t10_api_install_snapshot_with_lower_vote.rs
@@ -5,10 +5,10 @@ use anyhow::Result;
 use maplit::btreeset;
 use openraft::raft::AppendEntriesRequest;
 use openraft::raft::InstallSnapshotRequest;
+use openraft::storage::Snapshot;
+use openraft::storage::SnapshotMeta;
 use openraft::testing::log_id;
 use openraft::Config;
-use openraft::Snapshot;
-use openraft::SnapshotMeta;
 use openraft::Vote;
 
 use crate::fixtures::ut_harness;
